+++ app/config.py
@@ -33,7 +33,7 @@ class Settings(BaseSettings):
     MODE: ExecutionMode
     CHIA_ROOT: Path = Path("~/.chia/mainnet")
     CONFIG_PATH: Path = Path("climate_token/config/config.yaml")
-    SERVER_PORT: Optional[int]
+    SERVER_PORT: Optional[int] = 31999
 
     # Visible configs: configurable through config.yaml
     LOG_PATH: Path = Path("climate_token/log/debug.log")
@@ -67,7 +67,7 @@ class Settings(BaseSettings):
             cls._instance = get_settings()
         return cls._instance
 
-    @root_validator
+    @root_validator(skip_on_failure=True)
     def configure_port(cls, values: Dict[str, Any]) -> Dict[str, Any]:
         if values["MODE"] == ExecutionMode.REGISTRY:
             values["SERVER_PORT"] = values.get("CLIMATE_TOKEN_REGISTRY_PORT", ServerPort.CLIMATE_TOKEN_REGISTRY.value)