+++ app/core/climate_wallet/wallet.py
@@ -4,7 +4,6 @@ import dataclasses
 import time
 from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
 
-from chia_rs import AugSchemeMPL, G1Element, G2Element, PrivateKey
 from chia.consensus.constants import ConsensusConstants
 from chia.rpc.full_node_rpc_client import FullNodeRpcClient
 from chia.rpc.wallet_rpc_client import WalletRpcClient
@@ -12,7 +11,7 @@ from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.coin_record import CoinRecord
-from chia.types.coin_spend import CoinSpend
+from chia.types.coin_spend import CoinSpend, make_spend
 from chia.types.spend_bundle import SpendBundle, estimate_fees
 from chia.util.bech32m import bech32_decode, bech32_encode, convertbits
 from chia.util.ints import uint32, uint64
@@ -29,6 +28,8 @@ from chia.wallet.uncurried_puzzle import uncurry_puzzle
 from chia.wallet.util.compute_memos import compute_memos
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from chia.wallet.util.wallet_types import WalletType
+from chia.wallet.wallet_spend_bundle import WalletSpendBundle
+from chia_rs import AugSchemeMPL, G1Element, G2Element, PrivateKey
 
 from app.core.chialisp.gateway import create_gateway_puzzle, parse_gateway_spend
 from app.core.chialisp.tail import create_delegated_puzzle, create_tail_program
@@ -135,9 +136,9 @@ class ClimateWallet(ClimateWalletBase):
         return self.wallet_client is not None
 
     @property
-    def delegated_signatures(self) -> Dict[Tuple[bytes, bytes], G2Element]:
+    def delegated_signatures(self) -> Dict[Tuple[G1Element, bytes], G2Element]:
         return {
-            (bytes(self.root_public_key), message): signature
+            (self.root_public_key, message): signature
             for (
                 mode,
                 (message, signature),
@@ -188,7 +189,7 @@ class ClimateWallet(ClimateWalletBase):
         to_puzzle_hash: Optional[bytes32] = None,
         key_value_pairs: Optional[List[Tuple[Any, Any]]] = None,
         gateway_public_key: Optional[G1Element] = None,
-        public_key_to_secret_key: Optional[Dict[bytes, PrivateKey]] = None,
+        public_key_to_secret_key: Optional[Dict[G1Element, PrivateKey]] = None,
         allow_missing_signature: bool = False,
         wallet_id: int = 1,
     ) -> Dict[str, Any]:
@@ -224,35 +225,39 @@ class ClimateWallet(ClimateWalletBase):
             aggregated_signature=signature,
         )
 
-        transaction_records = await get_created_signed_transactions(
+        transactions = await get_created_signed_transactions(
             transaction_request=transaction_request,
             wallet_id=wallet_id,
             wallet_client=self.wallet_client,
         )
-        (first_transaction_record, *rest_transaction_records) = transaction_records
-        if first_transaction_record.spend_bundle is None:
-            raise ValueError("No spend bundle created!")
-        spend_bundle = SpendBundle.aggregate(
-            [
-                first_transaction_record.spend_bundle,
-                gateway_spend_bundle,
-            ]
-        )
-        first_transaction_record = dataclasses.replace(
-            first_transaction_record,
-            spend_bundle=spend_bundle,
-            additions=spend_bundle.additions(),
-            removals=spend_bundle.removals(),
-            type=uint32(CLIMATE_WALLET_INDEX + mode.to_int()),
-            name=spend_bundle.name(),
-            memos=list(compute_memos(spend_bundle).items()),
-        )
-        transaction_records = [first_transaction_record, *rest_transaction_records]
+        new_txs = []
+        for tx in transactions:
+            if unsigned_gateway_coin_spend.coin in tx.additions:
+                spend_bundle = WalletSpendBundle.aggregate(
+                    [gateway_spend_bundle] + ([] if tx.spend_bundle is None else [tx.spend_bundle])
+                )
+                additions = [
+                    add for add in tx.additions if add != unsigned_gateway_coin_spend.coin
+                ] + gateway_spend_bundle.additions()
+            else:
+                spend_bundle = WalletSpendBundle.aggregate(([] if tx.spend_bundle is None else [tx.spend_bundle]))
+                additions = tx.additions
+            removals = [rem for rem in tx.removals if rem not in additions]
+            new_tx = dataclasses.replace(
+                tx,
+                spend_bundle=spend_bundle,
+                additions=additions,
+                removals=removals,
+                type=uint32(CLIMATE_WALLET_INDEX + mode.to_int()),
+                name=spend_bundle.name(),
+                memos=list(compute_memos(spend_bundle).items()),
+            )
+            new_txs.append(new_tx)
 
         return {
-            "transaction_id": first_transaction_record.name,
-            "transaction_records": transaction_records,
-            "spend_bundle": spend_bundle,
+            "transaction_id": new_txs[0].name,
+            "transaction_records": new_txs,
+            "spend_bundle": SpendBundle.aggregate([tx.spend_bundle for tx in new_txs if tx.spend_bundle is not None]),
         }
 
     async def _create_client_transaction(
@@ -287,7 +292,6 @@ class ClimateWallet(ClimateWalletBase):
                 logger.info(f"    - {key}: {value}")
 
         transaction_request: TransactionRequest
-        transaction_records: List[TransactionRecord]
 
         # this is a hack to get inner puzzle hash for `origin_coin`
 
@@ -300,17 +304,17 @@ class ClimateWallet(ClimateWalletBase):
                     memos=[],
                 )
             ],
-            fee=0,
+            fee=uint64(0),
         )
-        transaction_records = await get_created_signed_transactions(
+        transactions = await get_created_signed_transactions(
             transaction_request=transaction_request,
             wallet_id=wallet_id,
             wallet_client=self.wallet_client,
         )
-        if len(transaction_records) != 1:
-            raise ValueError(f"Transaction record has unexpected length {len(transaction_records)}!")
+        if len(transactions) != 1:
+            raise ValueError(f"Transaction record has unexpected length {len(transactions)}!")
 
-        transaction_record = transaction_records[0]
+        transaction_record = transactions[0]
         if transaction_record.spend_bundle is None:
             raise ValueError("No spend bundle created!")
         coin_spend: CoinSpend = transaction_record.spend_bundle.coin_spends[0]
@@ -355,7 +359,7 @@ class ClimateWallet(ClimateWalletBase):
 
         gateway_secret_key: PrivateKey = self.mode_to_secret_key[mode]
         gateway_public_key: G1Element = self.mode_to_public_key[mode]
-        public_key_to_secret_key = {bytes(gateway_public_key): gateway_secret_key}
+        public_key_to_secret_key = {gateway_public_key: gateway_secret_key}
 
         coins: List[Coin] = await self.wallet_client.select_coins(
             amount=amount + fee,
@@ -458,15 +462,15 @@ class ClimateWallet(ClimateWalletBase):
             if puzzle_args is None:
                 continue
 
-            (_, asset_id, inner_puzzle) = puzzle_args
-            asset_id = asset_id.as_atom()
+            (_, asset_id_program, inner_puzzle) = puzzle_args
+            asset_id = asset_id_program.as_atom()
             inner_solution = solution.at("f")
 
             # check for gateway puzzle
             if inner_puzzle != gateway_puzzle:
                 continue
 
-            inner_coin_spend = CoinSpend(
+            inner_coin_spend = make_spend(
                 coin=coin,
                 puzzle_reveal=inner_puzzle,
                 solution=inner_solution,
@@ -523,7 +527,7 @@ class ClimateWallet(ClimateWalletBase):
             raise ValueError("No public keys provided for the registry!")
         gateway_secret_key: PrivateKey = self.mode_to_secret_key[mode]
         gateway_public_key: G1Element = self.mode_to_public_key[mode]
-        public_key_to_secret_key = {bytes(gateway_public_key): gateway_secret_key}
+        public_key_to_secret_key = {gateway_public_key: gateway_secret_key}
 
         (_, data) = bech32_decode(content, max_length=len(content))
         if data is None:
@@ -550,15 +554,16 @@ class ClimateWallet(ClimateWalletBase):
         if aggregated_signature == G2Element():
             raise ValueError("Invalid detokenization request!")
 
-        spend_bundle = SpendBundle.aggregate(
+        spend_bundle = WalletSpendBundle.aggregate(
             [
                 unsigned_spend_bundle,
-                SpendBundle(coin_spends=[], aggregated_signature=aggregated_signature),
+                WalletSpendBundle(coin_spends=[], aggregated_signature=aggregated_signature),
             ]
         )
         if gateway_coin_spend is None:
             raise ValueError("Invalid detokenization request: Could not find gateway coin spend!")
 
+        additions = spend_bundle.additions()
         transaction_record = TransactionRecord(
             confirmed_at_height=uint32(0),
             created_at_time=uint64(int(time.time())),
@@ -568,8 +573,8 @@ class ClimateWallet(ClimateWalletBase):
             confirmed=False,
             sent=uint32(0),
             spend_bundle=spend_bundle,
-            additions=spend_bundle.additions(),
-            removals=spend_bundle.removals(),
+            additions=additions,
+            removals=[rem for rem in spend_bundle.removals() if rem not in additions],
             wallet_id=uint32(wallet_id),
             sent_to=[],
             trade_id=None,
@@ -673,20 +678,20 @@ class ClimateObserverWallet(ClimateWalletBase):
             delegated_solution: Program = tail_solution.at("r")
             key_value_pairs: Program = delegated_solution.at("f")
 
-            metadata: Dict[bytes, bytes] = {}
+            metadata: Dict[str, str] = {}
             for key_value_pair in key_value_pairs.as_iter():
                 if (not key_value_pair.listp()) or (key_value_pair.at("r").listp()):
                     logger.warning(f"Coin {coin.name()} has incorrect metadata structure")
                     continue
 
-                key = key_value_pair.at("f").as_atom()
-                value = key_value_pair.at("r").as_atom()
+                key_bytes = key_value_pair.at("f").as_atom()
+                value_bytes = key_value_pair.at("r").as_atom()
 
-                key = key.decode()
+                key = key_bytes.decode()
                 if key in ["bp"]:
-                    value = f"0x{value.hex()}"
+                    value = f"0x{value_bytes.hex()}"
                 elif key in ["ba", "bn"]:
-                    value = value.decode()
+                    value = value_bytes.decode()
                 else:
                     raise ValueError("Unknown key '{key}'!")