+++ app/crud/chia.py
@@ -6,10 +6,10 @@ from typing import Any, Dict, List, Optional
 from urllib.parse import urlencode, urlparse
 
 import requests
-from chia_rs import G1Element
 from chia.rpc.full_node_rpc_client import FullNodeRpcClient
 from chia.types.blockchain_format.coin import Coin
 from chia.types.coin_record import CoinRecord
+from chia_rs import G1Element
 from fastapi.encoders import jsonable_encoder
 
 from app import schemas
@@ -46,7 +46,7 @@ class ClimateWareHouseCrud(object):
         Returns:
             A list of all data retrieved from the paginated API.
         """
-        all_data = []
+        all_data: list[dict[str, Any]] = []
         page = 1
         limit = 10
 
@@ -57,16 +57,32 @@ class ClimateWareHouseCrud(object):
                 encoded_params = urlencode(params)
 
                 # Construct the URL
-                url = urlparse(f"{self.url}{path}?{encoded_params}")
+                url_obj = urlparse(f"{self.url}{path}?{encoded_params}")
+                url = url_obj.geturl()
 
-                response = requests.get(url.geturl(), headers=self._headers())
+                response = requests.get(url, headers=self._headers())
                 if response.status_code != requests.codes.ok:
                     logger.error(f"Request Url: {response.url} Error Message: {response.text}")
                     raise error_code.internal_server_error(message="API Call Failure")
 
                 data = response.json()
+                if data is None:
+                    # some cadt endpoints return null with no pagination info if no data is found
+                    # to prevent an infinite loop need to assume that there is no data matching
+                    # the search from this iteration on
+                    return all_data
 
-                all_data.extend(data["data"])  # Add data from the current page
+                try:
+                    if (
+                        data["page"] and (data["pageCount"] >= 0) and len(data["data"]) >= 0
+                    ):  # page count can be 0 (as of when this was written)
+                        all_data.extend(data["data"])  # Add data from the current page
+                    else:
+                        all_data.append(data)  # data was not paginated, append and return
+                        return all_data
+                except Exception:
+                    all_data.append(data)  # data was not paginated, append and return
+                    return all_data
 
                 if page >= data["pageCount"]:
                     break  # Exit loop if all pages have been processed
@@ -179,7 +195,7 @@ class ClimateWareHouseCrud(object):
                 warehouse_project_id = unit["issuance"]["warehouseProjectId"]
                 project = project_by_id[warehouse_project_id]
             except (KeyError, TypeError):
-                logger.warning(f"Can not get project by warehouse_project_id: {warehouse_project_id}")
+                logger.warning("Can not get project by warehouse_project_id")
                 continue
 
             org_metadata = metadata_by_id.get(unit_org_uid)